# Fix performance regression in SSL handshake