@wfurt 2022-03-02
Can know upfront if we are in handshake or not and do only one of the calls as appropriate? And possibly know if the call is supported at all via `static bool` ?  This fall-back logic still little bit worries me.