@rzikm 2022-03-03
> Can know upfront if we are in handshake or not and do only one of the calls as appropriate?

If I see correctly, the `CertificateValidationPal.GetRemoteCertificate` has 2 overloads, one called during TLS handshake and one after the handshake completes, so we can differentiate between those two situations. I will try it out.

> And possibly know if the call is supported at all via `static bool` ?

I see a few options:
- set a flag once we encounter first failure, i.e. check for `SEC_E_UNSUPPORTED_FUNCTION` inside `SSPIWrapper` and short-circuit all following calls to `SSPIWrapper.QueryContextAttributes_SECPKG_ATTR_REMOTE_CERT_CHAIN`
- put OS version check in `CertificateValidationPal.GetRemoteCertificate`

I am not sure I like either of them. Since we can differentiate between in-handshake and after-handshake calls, I think we should just leave it be. Win8 and newer support the function (I am not sure about Win7, the test that requires it is disabled there) and besides, until now we already called `QueryContextAttributes` at least once, so there will be no additional perf hit there.