@rzikm 2022-03-03
Updated benchmarks look promising

```ini
// * Summary *                                                                                                                                                                           
                                                                                                                                                                                         
BenchmarkDotNet=v0.13.1.1702-nightly, OS=Windows 11 (10.0.22000.493/21H2)                                                                                                                
Intel Core i9-10900K CPU 3.70GHz, 1 CPU, 20 logical and 10 physical cores                                                                                                                
.NET SDK=7.0.100-preview.3.22152.16                                                                                                                                                      
  [Host]     : .NET 7.0.0 (7.0.22.11609), X64 RyuJIT                                                                                                                                     
  Job-QSWYNC : .NET 7.0.0 (42.42.42.42424), X64 RyuJIT                                                                                                                                   
  Job-YFPROJ : .NET 7.0.0 (42.42.42.42424), X64 RyuJIT                                                                                                                                   
  Job-IHEQPI : .NET 7.0.0 (42.42.42.42424), X64 RyuJIT                                                                                                                                   
                                                                                                                                                                                         
PowerPlanMode=00000000-0000-0000-0000-000000000000  Arguments=/p:DebugType=portable,-bl:benchmarkdotnet.binlog  IterationTime=250.0000 ms                                                
MaxIterationCount=20  MinIterationCount=15  WarmupCount=1                                                                                                                                
                                                                                                                                                                                         
|                           Method |        Job |             Toolchain |     Mean |     Error |    StdDev |   Median |      Min |      Max | Ratio | RatioSD | Allocated | Alloc Ratio |
|--------------------------------- |----------- |---------------------- |---------:|----------:|----------:|---------:|---------:|---------:|------:|--------:|----------:|------------:|
| DefaultHandshakeContextIPv6Async | Job-QSWYNC |    \7.0.0\corerun.exe | 1.439 ms | 0.0178 ms | 0.0167 ms | 1.438 ms | 1.411 ms | 1.475 ms |  1.00 |    0.02 |   5.58 KB |        1.00 |
| DefaultHandshakeContextIPv6Async | Job-YFPROJ |     \main\corerun.exe | 1.489 ms | 0.0232 ms | 0.0217 ms | 1.488 ms | 1.455 ms | 1.519 ms |  1.03 |    0.03 |   5.58 KB |        1.00 |
| DefaultHandshakeContextIPv6Async | Job-IHEQPI | \reverted\corerun.exe | 1.446 ms | 0.0254 ms | 0.0237 ms | 1.444 ms | 1.416 ms | 1.502 ms |  1.00 |    0.00 |    5.6 KB |        1.00 |
```